+++ apisrv/pkg/data_access/esdao/search_query_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tainvecs/apisrv/pkg/data_access/esdao"
-	"github.com/tainvecs/apisrv/pkg/utils/strutils"
+	"github.com/tainvecs/garage/apisrv/pkg/data_access/esdao"
+	"github.com/tainvecs/garage/apisrv/pkg/utils/strutils"
 )
 
 func TestNewTermQuery(t *testing.T) {
