+++ apisrv/resources/es-news_category.schema.json
