+++ apisrv/resources/stopwords.en.txt
