+++ apisrv/scripts/es-news_category-set_up.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
 
-GO_API_SERVER_ROOT="$(dirname $(cd $(dirname $0) >/dev/null 2>&1; pwd -P;))"
+APISRV_ROOT="$(dirname $(cd $(dirname $0) >/dev/null 2>&1; pwd -P;))"
 
 
 # create index
 curl -XPUT "$ES_URL/$ES_NEWS_DOC_INDEX" \
      -H 'Content-Type: application/json' \
-     -d "@$GO_API_SERVER_ROOT/resources/es-news_category.schema.json"
+     -d "@$APISRV_ROOT/resources/es-news_category.schema.json"
 
 
 # add index aliases
@@ -30,6 +30,6 @@ curl -XPOST "$ES_URL/_aliases" -H "Content-Type: application/json" -d'
 # # index test docs
 # curl -XPOST "$ES_URL/$ES_NEWS_DOC_INDEX_ALIAS/_bulk" \
 #      -H 'Content-Type: application/json' \
-#      --data-binary "@$GO_API_SERVER_ROOT/data/test-news_category.es.json"
+#      --data-binary "@$APISRV_ROOT/data/test-news_category.es.json"
 
 # curl -XPOST "$ES_URL/_refresh?pretty"
