+++ apisrv/scripts/es-unit_test-set_up.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
 
-GO_API_SERVER_ROOT="$(dirname $(cd $(dirname $0) >/dev/null 2>&1; pwd -P;))"
+APISRV_ROOT="$(dirname $(cd $(dirname $0) >/dev/null 2>&1; pwd -P;))"
 
 
 # create index
 curl -XPUT "$ES_URL/$ES_UNIT_TEST_INDEX" \
      -H 'Content-Type: application/json' \
-     -d "@$GO_API_SERVER_ROOT/resources/es-unit_test.schema.json"
+     -d "@$APISRV_ROOT/resources/es-unit_test.schema.json"
 
 
 # add index aliases
